// gb-angular-frontend/src/app/interfaces/user.ts
export interface User {
  user_id: number;
  first_name: string;
  last_name: string;
  username: string;
  password: string;
  role: string;
  email: string;
  icon: string;
  status: string;
  address: string;
  city: string;
  country: string;
  post_code: string;
  phone: string;
  balance: number;
  holder_name: string;
  account_number: string;
  bank_name: string
  branch_name: string;
  creation_date: string;
  user_alias: string;
  hold_fund: number;
}